Update AdditionalTopics.rst
Those two lines wrap together in markdown. BTW, the latter command gives me an error:

bitcoind --data-dir=~/.bitcoin-test
: Cannot obtain a lock on data directory ~/.bitcoin. Bitcoin is probably already running.

Which it is! The point is I want to run it on test.

This also gives an error:

bitcoind --data-dir ~/.bitcoin-test
error: {"code":-32601,"message":"Method not found"}

I ran bitcoin-qt with the --datadir command and that seems to be working fine: "Bitcoin Core -Wallet [testnet]" Synchronizing with network...